shopping_cart/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.14.0
mysqlclient>=2.2.0
python-dotenv>=1.0.0
django-cors-headers>=4.3.0
requests>=2.31.0
drf-yasg>=1.21.7